+++ tests/drift_kinetic_push/CMakeLists.txt
@@ -6,13 +6,11 @@ set(tests
   drift_kinetic_push_ex3
   drift_kinetic_push_ex4
   drift_kinetic_push_ex5
-  drift_kinetic_push_ex8
+  drift_kinetic_interpolation_ex1
   drift_kinetic_grid_boris_ex1
   drift_kinetic_grid_boris_ex2
-  drift_kinetic_grid_ex1
-  drift_kinetic_grid_ex2
-  drift_kinetic_grid_ex3
-  drift_kinetic_grid_ex4
+  drift_kinetic_grid_boris_ex3
+  drift_kinetic_grid_boris_ex4
 )
 
 # In the given tests (Omega * dt) is constant
